+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 - `v4_18` feature flag for to the `gtk4` crate to enable newer APIs.
 - `show_truncated` property of label widget to be false by default.
+- `can_shrink` property of image widget to be false by default.
 
 ### Fixed